worker/requirements.txt
jsonschema==4.21.1
python-dotenv==1.0.1
pytest==8.0.0
